Add note about how packages are resolved (#222)
In the main section it says that using `mint run` will run the latest version of a package which is installed. Then the Mintfile section says that you can use `mint run tool-name` which is the same as above and I was worried that it would run the latest version. (so I added note to clarify that the version would be resolved from the Mintfile)

We've got multiple projects over here and are using mint to install / manage package tools on our CI machines.